Add unit tests for Params
